backend/requirements.txt
fastapi
uvicorn[standard]
kafka-python
sqlalchemy
psycopg2-binary
redis
celery
python-dotenv
requests
